commit : Fix docker-compose and other infra things. Fix seeder and data is getting populated. Fix port-mapping and application is working from container as well as local.
