+++ Services/Catalog/Catalog.API/Catalog.API.csproj
@@ -27,7 +27,7 @@
 
     <ItemGroup>
       <ProjectReference Include="..\Catalog.Application\Catalog.Application.csproj" />
-      <ProjectReference Include="..\Catalog.Infrastrucure\Catalog.Infrastrucure.csproj" />
+      <ProjectReference Include="..\Catalog.Infrastructure\Catalog.Infrastructure.csproj" />
     </ItemGroup>
 
 </Project>
