+++ Services/Catalog/Catalog.API/appsettings.json
@@ -9,11 +9,11 @@
   
   "DatabaseSettings": {
     "MongoCatalog": {
-      "ConnectionStrings" : "mongodb://localhost:27017/?readPreference=primary&ssl=false",
+      "ConnectionString" : "mongodb://127.0.0.1:27017/?readPreference=primary&ssl=false",
       "DatabaseName" : "ProductCatalogDb",
-      "CollectionName" : "Products",
-      "BrandsCollection" : "Brands",
-      "TypesCollection" : "Types"
+      "ProductCollectionName" : "Products",
+      "ProductBrandCollectionName" : "Brands",
+      "ProductTypesCollectionName" : "Types"
     }
   }
 }
