+++ Services/Catalog/Catalog.Infrastructure/Class1.cs
