--- Services/Catalog/Catalog.Infrastrucure/Data/SeedData/types.json
@@ -1,22 +1,0 @@
-[
-  { "Id": "651a1a1b2d3a4b5c6d7e8f01", "ProductTypeName": "Shoes" },
-  { "Id": "651a1a1b2d3a4b5c6d7e8f02", "ProductTypeName": "T-Shirts" },
-  { "Id": "651a1a1b2d3a4b5c6d7e8f03", "ProductTypeName": "Jeans" },
-  { "Id": "651a1a1b2d3a4b5c6d7e8f04", "ProductTypeName": "Jackets" },
-  { "Id": "651a1a1b2d3a4b5c6d7e8f05", "ProductTypeName": "Sweatshirts" },
-  { "Id": "651a1a1b2d3a4b5c6d7e8f06", "ProductTypeName": "Shorts" },
-  { "Id": "651a1a1b2d3a4b5c6d7e8f07", "ProductTypeName": "Dresses" },
-  { "Id": "651a1a1b2d3a4b5c6d7e8f08", "ProductTypeName": "Skirts" },
-  { "Id": "651a1a1b2d3a4b5c6d7e8f09", "ProductTypeName": "Suits" },
-  { "Id": "651a1a1b2d3a4b5c6d7e8f0a", "ProductTypeName": "Blazers" },
-  { "Id": "651a1a1b2d3a4b5c6d7e8f0b", "ProductTypeName": "Sweaters" },
-  { "Id": "651a1a1b2d3a4b5c6d7e8f0c", "ProductTypeName": "Hoodies" },
-  { "Id": "651a1a1b2d3a4b5c6d7e8f0d", "ProductTypeName": "Polos" },
-  { "Id": "651a1a1b2d3a4b5c6d7e8f0e", "ProductTypeName": "Socks" },
-  { "Id": "651a1a1b2d3a4b5c6d7e8f0f", "ProductTypeName": "Caps" },
-  { "Id": "651a1a1b2d3a4b5c6d7e8f10", "ProductTypeName": "Bags" },
-  { "Id": "651a1a1b2d3a4b5c6d7e8f11", "ProductTypeName": "Belts" },
-  { "Id": "651a1a1b2d3a4b5c6d7e8f12", "ProductTypeName": "Watches" },
-  { "Id": "651a1a1b2d3a4b5c6d7e8f13", "ProductTypeName": "Glasses" },
-  { "Id": "651a1a1b2d3a4b5c6d7e8f14", "ProductTypeName": "Jewelry" }
-]
