--- Services/Catalog/Catalog.Infrastrucure/appsettings.json
@@ -1,5 +1,0 @@
-{
-  "ConnectionStrings": {
-    "Catalog": "mongodb://localhost:27017/CatalogDb"
-  }
-}
