+++ docker-compose.yml
@@ -1,0 +1,18 @@
+version: '3.4'
+
+services:
+  catalogdb:
+    image: mongo
+    ports:
+      - "27017:27017"
+    
+  catalog.api:
+    image: ${DOCKER_REGISTRY-}catalogapi
+    build:
+      context: .
+      dockerfile: Services/Catalog/Catalog.API/Dockerfile
+    ports:
+      - "9000:8080"
+
+volumes:
+  mongo_data:
